A server side file that identifies the id of the most recent set of art and where that art can be downloaded. Also indicates the expiration date (in UNIX timestamp) of the most recent set.
Local installations of the extension check this file upon their local copy of the art expiring. If the image_set.xml's id number is higher than the local set, the new set of art is downloaded and the expiration date is cached (so that the local machine knows when to start checking for new art).